colour validation added
