+++ lib/shapes.js
