+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "inquirer": "^8.2.4",
-    "inquirer-maxlength-input-prompt": "^1.0.2"
+    "inquirer-maxlength-input-prompt": "^1.0.2",
+    "validate-color": "^2.2.4"
   }
 }
